:art: 뷰모델 리팩토링
