+++ feature/mypage/build.gradle.kts
@@ -77,6 +77,7 @@ dependencies {
     ksp(libs.hilt.compiler)
     ksp(libs.androidx.hilt.complier)
     implementation(libs.androidx.hilt.navigation)
+    implementation(libs.androidx.hilt.navigation.compose)
 
     // FCM
     // ✅ Firebase (BOM + messaging-ktx)
